#35: Implement exit notification and remove process exit from shutdown
The shutdown message is currently not processed properly because VS Code
prematurely kills the Java Language Server and we don't have a chance
to finish the application thread and shut down the workspace.

This commit also monitors the parent processId to make sure we shut down
when the parent process died.