Adds quick fix for unnecessary cast
Fixes issue #165

To test you need to add the preference "org.eclipse.jdt.core.compiler.problem.unnecessaryTypeCheck=warning" to the org.eclipse.jdt.core.prefs settings file.
This will enable the warning to be shown in the ide.

And then refer to the test in this PR to use to re-create it.

Signed-off-by: Nikolas Komonen <nikolaskomonen@gmail.com>