Promote end-user usage instruction more in the README
The README was rather developer focused (and only targeting eclipse
users at that)

This puts the focus more on end-user usage.

The main motivation for this is to enable other editors or plugins to
point users to the installation instructions within this repository to
focus their own efforts on explaining how to configure the
editor/plugin.

Signed-off-by: Mathias Fussenegger <f.mathias@zignar.net>