Automatically add Protobuf output source directories to classpath
A Gradle init script is added as a patch to automatically add Protobuf
output source directories to classpath. The script works for Gradle
'com.google.protobuf' plugin 0.8.4 or higher.

The init script will be extracted to a cache folder.

As a side effect, the Protobuf output source directories will be created
beforehand, due to an issue of Buildship.
See: https://github.com/eclipse/buildship/issues/1196.

Signed-off-by: sheche <sheche@microsoft.com>