Support adding imports for postfix completion
Different from Eclipse just apply the text edits directly after a completion
item is committed. In JDT.LS, we have to calculate the additional text edits
first and return them to the client. And then wait the client to apply the
edits and send the document change event back to the server. So the implementation
is different from JDT.Core.

The import statement to be added will be calculated when resolvers are resolving
the variables. The result will be cached in a map organized by the template name.

Signed-off-by: Sheng Chen <sheche@microsoft.com>