Display the postfix completion at the bottom of the list
This PR put the postfix templates to the bottom of the completion list,
which is also the default behavior of other popular Java IDEs.

Previously, a text edit is used to replace the expression. To make it work,
another filter text has to be set to match the expression. A side effect
of this is that it always places the postfix template at the top of the list
because the filter text is matched.

In this change, the insert text is directly used, so the filter text is
not needed anymore. Besides, another additional text edit will be responsible
to remove the code to be replaced.

Signed-off-by: Sheng Chen <sheche@microsoft.com>