Log user friendly error if client does not support _java.reloadBundles.command
The result of `executeClientCommand` could be a map in the LSP
`ResponseError` format.
Previously this resulted in a ClassCastException which got logged:

	class com.google.gson.internal.LinkedTreeMap cannot be cast to class java.util.ArrayList (com.google.gson.internal.LinkedTreeMap

This gave users the impression that there's a bug.
This changes the logic to instead log the message returned by the
client.

In neovim for example it results in:

	Command _java.reloadBundles.command not supported on client