Make Start Up Script compatible with older Python
Previously the start up script created to be friendly to package
managers used the `BooleanOptionalAction` in its argument parser. This
feature is only available in Python 3.9 and later. Since the script
shebang in jdtls indicates only Python3, this change updates the script
to conform with that compatability.

This fix was suggested by tmmvn in [issue
2153](https://github.com/eclipse/eclipse.jdt.ls/issues/2153). This
     change implements that fix.