Fix duplicated content in test class
Using `createCompilationUnit` then `createType` results in duplicated
content in the file. This introduces some extra unrelated problems in
the file, with both ECJ and Javac, that are not in the scope of this
test.
Supporting this extra duplication story is more extra work for the test
to pass with Javac. So we fix the test to make it work out of the box
with Javac too.