Improve rename tests so they can work with Javac too
The current code being tested has syntax errors. Those errors prevents
Javac from properly setting up bindings (and it's not obvious that Javac
is more wrong than ECJ, so there are little chance Javac gets changed to
behave like ECJ).
Those syntax errors do not seem so relevant for the case being covered.
So we get rid of them to focus on the "correct" case and allow Javac to
work with it.